requirements.txt
Flask==2.1.1
requests==2.26.0
Werkzeug==2.3.7
flask_cors==3.0.10
dashscope==1.13.6
mem0ai
chromadb
